Refactor user/group search into a single component
